Teh warnings are gone
git-svn-id: file:///home/skmp/Desktop/public_svn/nullDC@156 d5fdb02f-ce0d-0410-91b1-fe443107cc8c

+++ trunk/nulldc/nullDC/dc/aica/aica_if.cpp
@@ -39,7 +39,7 @@ u32 ReadMem_aica_rtc(u32 addr,u32 sz)
 		rawtime+=24*3600;//add an hour if dst (maby rtc has a reg for that ? *watch* and add it if yes :)
 
 
-	u32 RTC=0x5bfc8900 + rawtime;// add delta to known dc time
+	u32 RTC=0x5bfc8900 + (u32)rawtime;// add delta to known dc time
 
 	switch( addr & 0xFF )
 	{
